+++ CHANGELOG.md
@@ -6,8 +6,8 @@ CHANGELOG
 
 * Updated `MaxMind\MinFraud\ReportTransaction` to make the `ip_address`
   parameter optional. Now the `tag` and at least one of the following
-  parameters must be supplied: `ip_address`, `maxmind_id`, `minfraud_id`,
-  `transaction_id`.
+  parameters must be supplied: `ipAddress`, `maxmindId`, `minfraudId`,
+  `transactionId`.
 
 3.0.1 (2024-05-02)
 ------------------